Extract json_selection/graphql.rs from parser.rs.
